requirements.txt
Pillow == 9.1.1
fastapi == 0.78.0
pydantic == 1.9.1
numpy == 1.25.2
uvicorn == 0.17.6
boto3 == 1.24.9
faiss-cpu == 1.7.4
torchvision == 0.15.2
torch == 2.0.1
transformers
python-multipart

